add license and rename to mocks
